added larger image compatability with mulitple meshes
