+++ Bucharest/Assets/Resources/MapChunk.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 660febd96e65a3348aab856546dd5ed9
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
