--- Bucharest/Assets/Scripts/Editor Scripts/MapGeneratorEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c5911bb627743744a9258ed579df729a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
