+++ Bucharest/Assets/Scripts/MapChunk.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 49a0b65d1b2f1b647a280fbb13e9c6b3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
