+++ Bucharest/Assets/Scripts/MapDistributer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fe5206521b0ea0d4e9848adba2c0f83e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
